Move gas estimate to signer when sending txs via provider
Update test comment on account nonce
